feat(timing): add (theoretically) nanosecond precision timing for compute loops

+++ templates/base1.py
@@ -100,7 +100,7 @@ int main(int argc, char **argv) {
 		}
   	}
 
-	printf("\\"compute\\": %.1f, \\"code\\": 0}\\n", ((double) (end - start)));
+	printf("\\"compute\\": %d, \\"code\\": 0}\\n", ((int) (end - start)));
 
 	free(input);
 	free(output);
